backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
numpy==2.0.1
PyJWT[crypto]==2.10.1
